requirements.txt
streamlit>=1.28.0
openai>=1.3.0
streamlit-authenticator>=0.2.3
pandas>=1.5.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
openpyxl>=3.1.0
bcrypt>=4.0.0
pyyaml>=6.0.0 